+++ src/pytorch_lightning/loggers/wandb.py
@@ -273,6 +273,7 @@ class WandbLogger(Logger):
 
         prefix: A string to put at the beginning of metric keys.
         experiment: WandB experiment object. Automatically set when creating a run.
+        checkpoint_name: Name of the model checkpoint artifact being logged.
         \**kwargs: Arguments passed to :func:`wandb.init` like `entity`, `group`, `tags`, etc.
 
     Raises:
@@ -298,6 +299,7 @@ class WandbLogger(Logger):
         log_model: Union[str, bool] = False,
         experiment: Union[Run, RunDisabled, None] = None,
         prefix: str = "",
+        checkpoint_name: Optional[str] = None,
         **kwargs: Any,
     ) -> None:
         if wandb is None:
@@ -353,6 +355,8 @@ class WandbLogger(Logger):
         if _WANDB_GREATER_EQUAL_0_12_10:
             wandb.require("service")
             _ = self.experiment
+
+        self._checkpoint_name = checkpoint_name
 
     def __getstate__(self) -> Dict[str, Any]:
         state = self.__dict__.copy()
@@ -591,7 +595,9 @@ class WandbLogger(Logger):
                 if _WANDB_GREATER_EQUAL_0_10_22
                 else None
             )
-            artifact = wandb.Artifact(name=f"model-{self.experiment.id}", type="model", metadata=metadata)
+            if not self._checkpoint_name:
+                self._checkpoint_name = f"model-{self.experiment.id}"
+            artifact = wandb.Artifact(name=self._checkpoint_name, type="model", metadata=metadata)
             artifact.add_file(p, name="model.ckpt")
             self.experiment.log_artifact(artifact, aliases=[tag])
             # remember logged models - timestamp needed in case filename didn't change (lastkckpt or custom name)